+++ src/components/ui/form/SearchableSettingRow.tsx
@@ -5,14 +5,24 @@ interface SearchableSettingRowProps {
   searchTerm?: string;
   searchMatches?: string[];
   children?: ReactNode;
+  title?: string;
+  indent?: number;
+  isCheckbox?: boolean;
 }
 
 export const SearchableSettingRow: FC<SearchableSettingRowProps> = ({
   searchTerm,
   searchMatches,
   children,
+  title,
+  indent,
+  isCheckbox,
 }) => {
-  const component = <SettingRow>{children}</SettingRow>;
+  const component = (
+    <SettingRow title={title} $indent={indent} $isCheckbox={isCheckbox}>
+      {children}
+    </SettingRow>
+  );
   if (!searchTerm || !searchMatches || searchTerm.length === 0) {
     return component;
   }